+++ tests/Helpers/Pdo/MysqlPdo.php
@@ -21,6 +21,8 @@ class MysqlPdo extends PDO
         if (getenv('PHOENIX_MYSQL_CHARSET')) {
             $dsn .= ';charset=' . getenv('PHOENIX_MYSQL_CHARSET');
         }
-        parent::__construct($dsn, getenv('PHOENIX_MYSQL_USERNAME'), getenv('PHOENIX_MYSQL_PASSWORD'));
+        parent::__construct($dsn, getenv('PHOENIX_MYSQL_USERNAME'), getenv('PHOENIX_MYSQL_PASSWORD'), [
+            PDO::ATTR_ERRMODE => PDO::ERRMODE_SILENT
+        ]);
     }
 }