+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 #### Added
 - support for PHP 8.0 - `$pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_SILENT);`
+- set autoincrement for new table and also for existing table
 
 ### [1.3.0] - 2020-04-16
 #### Added